test(teams): move catalog coverage to its sibling test module
Keep the catalog regressions together while meeting the 500-line inline-test module limit after adding missing-definition coverage.
